+++ src/msgpack/vrefbuffer.hpp
@@ -40,7 +40,7 @@ public:
 	}
 
 public:
-	void write(const char* buf, unsigned int len)
+	void write(const char* buf, size_t len)
 	{
 		if(len < base::ref_size) {
 			append_copy(buf, len);
